requirements.txt
strands-agents
strands-agents-tools
streamlit>=1.30.0